LC-separate-black-and-white-balls [Runtime: N/A; Memory: N/A]

+++ greedy/separate-black-and-white-balls.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     long long minimumSteps(string s) {
         int n = s.length();
-        int count = 0;
-        int ones = 0;
-        int i = 0;
+        long long count = 0;
+        long long ones = 0;
+        long long i = 0;
         while(i < n){
             if(s[i] == '1'){
                 ones+=1;
